+++ lib/features/domain/usecases/get_expense_in_month_use_case.dart
@@ -3,13 +3,13 @@ import 'package:expense_app/core/error/failure.dart';
 import 'package:expense_app/core/usecase/usecase.dart';
 import 'package:expense_app/features/domain/repositories/expense_repository.dart';
 
-class GetExpenseInMonthUseCase extends UseCase<int, GetExpenseInMonthUseCaseParams?> {
+class GetExpenseInMonthUseCase extends UseCase<double, GetExpenseInMonthUseCaseParams?> {
 
   final ExpenseRepository repository;
   GetExpenseInMonthUseCase({required this.repository});
 
   @override
-  Future<Either<Failure, int>> call(GetExpenseInMonthUseCaseParams? params) async {
+  Future<Either<Failure, double>> call(GetExpenseInMonthUseCaseParams? params) async {
     return await repository.getExpenseInMonth(params?.fromDate, params?.untilDate);
   }
