+++ lib/features/domain/usecases/get_total_funds_use_case.dart
@@ -3,14 +3,14 @@ import 'package:expense_app/core/error/failure.dart';
 import 'package:expense_app/core/usecase/usecase.dart';
 import 'package:expense_app/features/domain/repositories/expense_repository.dart';
 
-class GetTotalFundsUseCase extends UseCase<int, GetTotalFundsUseCaseParams> {
+class GetTotalFundsUseCase extends UseCase<double, GetTotalFundsUseCaseParams> {
 
   final ExpenseRepository repository;
 
   GetTotalFundsUseCase({required this.repository});
 
   @override
-  Future<Either<Failure, int>> call(GetTotalFundsUseCaseParams params) async {
+  Future<Either<Failure, double>> call(GetTotalFundsUseCaseParams params) async {
     return await repository.getTotalFunds(params.fromDate, params.untilDate);
   }
